requirements.txt
pytest-playwright
lxml
openai
dacite
bs4